Many files:   dosio.c: New file to do DOS/BIOS disk accesses.   namei.c (open_namei): Make pathlen be of type size_t.   llseek.c: Always #include stdlib.h since it's need to define   	size_t.   io.h: Use errcode_t for magic numbers.   icount.c (get_icount_el), dupfs.c (ext2fs_dup_handle), dblist.c   	(dir_block_cmp): Use size_t where appropriate.   read_bb.c (ext2fs_read_bb_inode), cmp_bitmaps.c   	(ext2fs_compare_inode_bitmap): Use blk_t, ino_t and size_t   	where appropriate.   closefs.c (ext2fs_flush): Use dgrp_t instead of int where   	appropriate.   openfs.c (ext2fs_open), check_desc.c (ext2fs_check_desc): Use blk_t   	instead of int where 	appropriate.   rw_bitmaps.c (read_bitmaps), irel_ma.c, inode.c (ext2fs_write_inode),   	initialize.c (ext2fs_initialize): brel_ma.c: Fix to make be   	16-bit safe.   link.c (ext2fs_link), unlink.c (ext2fs_unlink), lookup.c (lookup_proc),   	ismounted.c (ext2fs_check_if_mounted), block.c (xlate_func):   	Add #pragma argsused for Turbo C.

+++ lib/ext2fs/brel_ma.c
@@ -3,6 +3,9 @@
  * 
  * Copyright (C) 1996, 1997 Theodore Ts'o.
  *
+ * TODO: rewrite to not use a direct array!!!  (Fortunately this
+ * module isn't really used yet.)
+ *
  * %Begin-Header%
  * This file may be redistributed under the terms of the GNU Public
  * License.
@@ -72,7 +75,8 @@ errcode_t ext2fs_brel_memarray_create(char *name, blk_t max_block,
 	memset(ma, 0, sizeof(struct brel_ma));
 	brel->private = ma;
 	
-	size = sizeof(struct ext2_block_relocate_entry) * (max_block+1);
+	size = (size_t) (sizeof(struct ext2_block_relocate_entry) *
+			 (max_block+1));
 	ma->entries = malloc(size);
 	if (!ma->entries)
 		goto errout;
@@ -106,7 +110,7 @@ static errcode_t bma_put(ext2_brel brel, blk_t old,
 	ma = brel->private;
 	if (old > ma->max_block)
 		return EINVAL;
-	ma->entries[old] = *ent;
+	ma->entries[(unsigned)old] = *ent;
 	return 0;
 }
 
@@ -118,7 +122,7 @@ static errcode_t bma_get(ext2_brel brel, blk_t old,
 	ma = brel->private;
 	if (old > ma->max_block)
 		return EINVAL;
-	if (ma->entries[old].new == 0)
+	if (ma->entries[(unsigned)old].new == 0)
 		return ENOENT;
 	*ent = ma->entries[old];
 	return 0;
@@ -137,10 +141,10 @@ static errcode_t bma_next(ext2_brel brel, blk_t *old,
 
 	ma = brel->private;
 	while (++brel->current < ma->max_block) {
-		if (ma->entries[brel->current].new == 0)
+		if (ma->entries[(unsigned)brel->current].new == 0)
 			continue;
 		*old = brel->current;
-		*ent = ma->entries[brel->current];
+		*ent = ma->entries[(unsigned)brel->current];
 		return 0;
 	}
 	*old = 0;
@@ -154,10 +158,10 @@ static errcode_t bma_move(ext2_brel brel, blk_t old, blk_t new)
 	ma = brel->private;
 	if ((old > ma->max_block) || (new > ma->max_block))
 		return EINVAL;
-	if (ma->entries[old].new == 0)
+	if (ma->entries[(unsigned)old].new == 0)
 		return ENOENT;
-	ma->entries[new] = ma->entries[old];
-	ma->entries[old].new = 0;
+	ma->entries[(unsigned)new] = ma->entries[old];
+	ma->entries[(unsigned)old].new = 0;
 	return 0;
 }
 
@@ -168,9 +172,9 @@ static errcode_t bma_delete(ext2_brel brel, blk_t old)
 	ma = brel->private;
 	if (old > ma->max_block)
 		return EINVAL;
-	if (ma->entries[old].new == 0)
+	if (ma->entries[(unsigned)old].new == 0)
 		return ENOENT;
-	ma->entries[old].new = 0;
+	ma->entries[(unsigned)old].new = 0;
 	return 0;
 }
